+++ src/components/layouts/Layout3.jsx
@@ -10,10 +10,12 @@ function Layout({ Conteudo, conteudoProps }) {
   return (
     <>
       <Header />
-      <NavBar />
-      <div className="flex justify-center">
-        {Conteudo ? <Conteudo {...conteudoProps} /> : null}
-        <TagsPopulares tags={tags} />
+      <div className="max-w-6xl mx-auto px-0 sm:px-0 md:px-4 lg:px-4">
+        <NavBar />
+        <div className="flex gap-4 justify-center ml-0 sm:ml-0 md:ml-4 lg:ml-4 mb-8">
+          {Conteudo ? <Conteudo {...conteudoProps} /> : null}
+          <TagsPopulares tags={tags} />
+        </div>
       </div>
     </>
   );