+++ src/components/layouts/Layout4.jsx
@@ -1,18 +1,16 @@
 import LayoutBase from './Layout3';
 import PerfilBox from './PerfilBox';
-import fotoDePerfil from '../../assets/fotoDePerfil.jpg';
 
 //Layout Padrão com Header completo, Tags Populares e box de Perfil
-// É possível definir o nome do usuário exibido, seu username e sua foto de perfil através dos parâmetros
-function Layout({ nomeUsuario = 'Teste', usernameUsuario = '@teste' }) {
+// É possível definir o usuário exibido
+function Layout({ usuario, loadingUsuario }) {
   return (
     <>
       <LayoutBase
         Conteudo={PerfilBox}
         conteudoProps={{
-          nomeUsuario,
-          usernameUsuario,
-          fotoDePerfil,
+          usuario,
+          loadingUsuario,
         }}
       />
     </>