+++ src/router/routes.jsx
@@ -7,11 +7,13 @@ import Layout4 from '../components/layouts/Layout4';
 import RecuperarSenha from '../pages/RecuperarSenha';
 import NovaSenha from '../pages/NovaSenha';
 import CriarConta from '../pages/CriarConta';
+import Perfil from '../pages/perfil/Perfil';
 
 export default function RoutesApp() {
   return (
     <Routes>
       <Route path="/" element={<Navigate to="/login" replace />} />
+
       <Route
         path="/login"
         element={
@@ -20,6 +22,7 @@ export default function RoutesApp() {
           </PublicRoute>
         }
       />
+
       <Route
         path="/feed"
         element={
@@ -28,10 +31,20 @@ export default function RoutesApp() {
           </PrivateRoute>
         }
       />
-      <Route path="*" element={<NaoEncontrado />} />
+
+      <Route
+        path="/perfil/:username"
+        element={
+          <PrivateRoute>
+            <Perfil />
+          </PrivateRoute>
+        }
+      />
+
       <Route path="/recuperar-senha" element={<RecuperarSenha />} />
       <Route path="/alterar-senha" element={<NovaSenha />} />
       <Route path="/cadastrar-usuario" element={<CriarConta />} />
+      <Route path="*" element={<NaoEncontrado />} />
     </Routes>
   );
 }