+++ CHANGELOG.md
@@ -42,3 +42,12 @@
 ### Adicionado
 
 - Modal de Avaliação, criar e editar avaliação
+
+## US5.5 - PR 9
+
+### Adicionando
+
+- Box de avaliação responsivo
+- Interatividade com avaliação (curtir e acessar perfil de autor)
+- Exibição de avaliações de Usuário em seu Perfil
+- Exibição do perfil de outros usuários